BEO-11463 add ppv to campaign + apply campaign with coupon
* Add channel condition to campaign promotion

* Add channel condition to campaign promotion

* Add channel condition to campaign promotion

* fix response

* FileTYpeCondition

* no message

* change promotion to base

* Add KalturaCouponPromotion

* change promotion to base

* update conditionValidation

* update conditionValidation

* change promotion to base

* Condition

* no message

* fix channel validation

* Add ordering

* fix validation

* Adding filter and sorting changes

* implementation

* add permissions

* rename

* fix permission

* refactoring

* refactoring

* no message

* refactoring

* refactoring

* fix stateIn

* refactoring

* refactoring

* refactoring

* refactoring

* no message

* no message

* fix ordering

* change promotion to base

* Adding currencyId

* Adding currencyId

* support campaigns for ppv (#1394)

* WIP SUPPORT CAMPAIGN IN PPV

* run reflector

* return OriginalPrice

* wip coupon campaign

* wip coupon for campaign

* add mediaid and filetypeidsto getvalid campaign

* reflector

* fix small

* move GetValidCampaign into GetMediaFileFinalPrice

* implement todos

* try to fix build

* fix BEO-12235 get real discountModule

* add default

* wip set country code in price

* discountditails fix

* fix DiscountModule in ppv

* sey currency in subscription init

* BEO-12325 edit campaign crud errors

* update tests


* fix sorting

* fix sorting

* BEO-12495 fix filtering by state

* run reflector

* comment CampaignManagerTests

* comment tests

* fix BEO-12495

* fix ut

* fix min validation

* fix updater id on internal update

* batch scope

* improve BatchCampaignConditionScope

* fix min validations

* fix

+++ Core/WebAPI/Managers/Scheme/SchemeInput.cs
@@ -193,7 +193,7 @@ namespace WebAPI.Managers.Scheme
         {
             if (MinInteger > int.MinValue)
             {
-                long lValue = (long)Convert.ChangeType(value, typeof(long));
+                int lValue = (int)Convert.ChangeType(value, typeof(int));
                 if (lValue < MinInteger)
                     throw new BadRequestException(BadRequestException.ARGUMENT_MIN_VALUE_CROSSED, name, MinInteger.ToString());
             }
@@ -213,7 +213,7 @@ namespace WebAPI.Managers.Scheme
         {
             if (MinLong > long.MinValue)
             {
-                int lValue = (int)Convert.ChangeType(value, typeof(int));
+                long lValue = (long)Convert.ChangeType(value, typeof(long));
                 if (lValue < MinLong)
                     throw new BadRequestException(BadRequestException.ARGUMENT_MIN_VALUE_CROSSED, name, MinLong.ToString());
             }
